Move the error handler out
